// api/typeuser.ts
export class Typeuser {
    id: any;
    version: any;
    code: any;
    libelle: any;
    P_codeStation: any;
    P_codeExercice: any;
    j_export: any;
    j_ddm: any;
    j_codStation: any;
    j_codeUser: any;
    j_operation: any;
    isDefault: any;
    affich_Prix_Achat: any;
    modif_Prix_Enc: any;
    technicien: any;
    sousTraitant: any;
    administrateur: any;
    superAdministrateur: any;
    acceptReglementSansDoc: any;
    AjoutModifArticle: any;
    apporteur: any;
    commercial: any;
    istechspec: any;
    timesheet: any;
    developpeur: any;
    aff_Multistock: any;
}